requirements.txt
Django>=5.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
scikit-learn
transformers
torch
nltk
requests
httplib2
